requirements.txt
discord.py==2.3.2
aiohttp==3.8.6
aiosqlite==0.19.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
validators==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
colorlog==6.7.0
cryptography==41.0.7
Pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.1.2